Minor cast fixes and confighandler header fix
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@452 37977431-3df6-0310-b722-df95706aa16b

+++ rts/CobInstance.cpp
@@ -844,10 +844,10 @@ int CCobInstance::GetUnitVal(int val, int p1, int p2, int p3, int p4)
 	default:
 		info->AddLine("CobError: Unknown get constant %d", val);
 	case VETERAN_LEVEL: 
-		return 100*unit->experience;
+		return (int)(100*unit->experience);
 	case CURRENT_SPEED:
 		if (unit->moveType)
-			return unit->speed.Length()*SCALE;
+			return (int)(unit->speed.Length()*SCALE);
 		return 0;
 	case ON_ROAD:
 		return 0;
